tela para adicionar um grupo de permissão

+++ views/permissions.php
@@ -7,8 +7,25 @@
 
 <div class="tabcontent">
     <div class="tabbody">
-        GRUPOS DE PERMISSOES
+        <a class="button" href="<?= BASE_URL ?>/permissions/add_group">Adicionar Grupo de Permissão</a>
+
+        <table width="100%">
+            <tr>
+                <th>Nome do Grupo de Permissões</th>
+                <th width="160">Ações</th>
+            </tr>
+            <?php foreach ($permission_groups_list as $p): ?>
+                <tr>
+                    <td><?= $p['name'] ?></td>
+                    <td>
+                        <a class="button button_small" href="<?= BASE_URL ?>/permissions/edit_group/<?= $p['id'] ?>">Editar</a>
+                        <a class="button button_small" href="<?= BASE_URL ?>/permissions/delete_group/<?= $p['id'] ?>" onclick="return confirm('Tem certeza que deseja excluir ?')">Excluir</a>
+                    </td>
+                </tr>
+            <?php endforeach; ?>
+        </table>
     </div>
+
 
     <div class="tabbody">
         <a class="button" href="<?= BASE_URL ?>/permissions/add">Adicionar Permissão</a>
@@ -21,7 +38,7 @@
             <?php foreach ($permission_list as $p): ?>
                 <tr>
                     <td><?= $p['name'] ?></td>
-                    <td><a class="button button_small" href="<?= BASE_URL ?>/permissions/delete/<?= $p['id'] ?>">Excluir</a></td>
+                    <td><a class="button button_small" href="<?= BASE_URL ?>/permissions/delete/<?= $p['id'] ?>" onclick="return confirm('Tem certeza que deseja excluir ?')">Excluir</a></td>
                 </tr>
             <?php endforeach; ?>
         </table>
